m-monitor: connect the monitor core on activation instead of module load time
We have multiple instances of the monitor plugin, but that's ok.
connect/disonnect will not do anything bad if called multiple times

We need to connect later so that the first connection is the one from main().
Otherwise, if there is a connection error, we will see the warning from the
monitor first.