properties: allow `a{sv}` variants in wp_properties_new_from_variant()
This is because the anatole engine converts lua tables into `a{sv}`
variants and we want an easy way to convert those to WpProperties.

When converting WpProperties for passing them to lua this is not an issue
because anatole is capable of converting any kind of dictionary into
a lua table.