scripts: intended-roles: Use both device and stream
When a stream gets added, we might have a situation where the device
does not yet exist to actually link on the basis of intended role. For
example this can be observed if no mic sources are present and then the
mic appears after a bluetooth profile switch to HFP.

To fix this, iterate over streams when a device connects in addition to
iterating over devices when a stream connects.